downloads negative examples

+++ loaders.py
@@ -26,7 +26,23 @@ def bbox(W, S, E, N):
         return readjson(fname)
     
     api = overpass.API()
-    response = api.get(f"node[\"highway\"=\"street_lamp\"]({S}, {W}, {N}, {E})", responseformat='json')
+    query = f"node[\"highway\"=\"street_lamp\"]({S}, {W}, {N}, {E})"
+    response = api.get(query, responseformat='json', verbosity='skel')
+    with open(fname, 'w') as json_file:
+        json.dump(response, json_file)
+    out = [(e['lat'], e['lon']) for e in response['elements']]
+    print("input len:", len(out))
+    return out
+
+def roads_bbox(W, S, E, N):
+    # same but road nodes instead of lamps
+    fname = f"./in/roads_bbox{mil(W)}_{mil(S)}_{mil(E)}_{mil(N)}.json"
+    if os.path.isfile(fname):
+        return readjson(fname)
+    
+    api = overpass.API()
+    query = f"(way[\"highway\"=\"primary\"]({S}, {W}, {N}, {E});way[\"highway\"=\"secondary\"]({S}, {W}, {N}, {E});way[\"highway\"=\"tertiary\"]({S}, {W}, {N}, {E});)->.ways;node(w.ways);"
+    response = api.get(query, responseformat='json', verbosity='skel')
     with open(fname, 'w') as json_file:
         json.dump(response, json_file)
     out = [(e['lat'], e['lon']) for e in response['elements']]
@@ -35,7 +51,10 @@ def bbox(W, S, E, N):
 
 def poly(bounds):
     # returns points enclosed by a polygon
+    # https://wiki.openstreetmap.org/wiki/Overpass_API/Overpass_QL#Polygon_evaluator
     pass
+
+
 
 if __name__ == "__main__":
     start = time.time()
